2.0: Refit migrations
