+++ updates/000001_create_posts.php
@@ -1,18 +1,15 @@
-<?php namespace RainLab\Blog\Updates;
+<?php
 
-use Schema;
+use October\Rain\Database\Schema\Blueprint;
 use October\Rain\Database\Updates\Migration;
 
-class CreatePostsTable extends Migration
+return new class extends Migration
 {
-
     public function up()
     {
-        Schema::create('rainlab_blog_posts', function($table)
-        {
-            $table->engine = 'InnoDB';
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->nullable()->index();
+        Schema::create('rainlab_blog_posts', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->nullable()->index();
             $table->string('title')->nullable();
             $table->string('slug')->index();
             $table->text('excerpt')->nullable();
@@ -20,6 +17,7 @@ class CreatePostsTable extends Migration
             $table->longText('content_html')->nullable();
             $table->timestamp('published_at')->nullable();
             $table->boolean('published')->default(false);
+            $table->mediumText('metadata')->nullable();
             $table->timestamps();
         });
     }
@@ -28,5 +26,4 @@ class CreatePostsTable extends Migration
     {
         Schema::dropIfExists('rainlab_blog_posts');
     }
-
-}
+};
