+++ updates/000002_create_categories.php
@@ -1,0 +1,28 @@
+<?php
+
+use October\Rain\Database\Schema\Blueprint;
+use October\Rain\Database\Updates\Migration;
+
+return new class extends Migration
+{
+    public function up()
+    {
+        Schema::create('rainlab_blog_categories', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->string('name')->nullable();
+            $table->string('slug')->nullable()->index();
+            $table->string('code')->nullable();
+            $table->text('description')->nullable();
+            $table->bigInteger('parent_id')->unsigned()->nullable()->index();
+            $table->integer('nest_left')->nullable();
+            $table->integer('nest_right')->nullable();
+            $table->integer('nest_depth')->nullable();
+            $table->timestamps();
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('rainlab_blog_categories');
+    }
+};
