--- updates/create_categories_table.php
@@ -1,41 +1,0 @@
-<?php namespace RainLab\Blog\Updates;
-
-use Schema;
-use October\Rain\Database\Updates\Migration;
-
-class CreateCategoriesTable extends Migration
-{
-
-    public function up()
-    {
-        Schema::create('rainlab_blog_categories', function($table)
-        {
-            $table->engine = 'InnoDB';
-            $table->increments('id');
-            $table->string('name')->nullable();
-            $table->string('slug')->nullable()->index();
-            $table->string('code')->nullable();
-            $table->text('description')->nullable();
-            $table->integer('parent_id')->unsigned()->index()->nullable();
-            $table->integer('nest_left')->nullable();
-            $table->integer('nest_right')->nullable();
-            $table->integer('nest_depth')->nullable();
-            $table->timestamps();
-        });
-
-        Schema::create('rainlab_blog_posts_categories', function($table)
-        {
-            $table->engine = 'InnoDB';
-            $table->integer('post_id')->unsigned();
-            $table->integer('category_id')->unsigned();
-            $table->primary(['post_id', 'category_id']);
-        });
-    }
-
-    public function down()
-    {
-        Schema::dropIfExists('rainlab_blog_categories');
-        Schema::dropIfExists('rainlab_blog_posts_categories');
-    }
-
-}
