+++ updates/version.yaml
@@ -1,8 +1,9 @@
 v1.0.1:
     - Initialize plugin.
-    - create_posts_table.php
-    - create_categories_table.php
-    - seed_all_tables.php
+    - 000001_create_posts.php
+    - 000002_create_categories.php
+    - 000003_create_posts_categories.php
+    - seed_tables.php
 v1.0.2: Added the processed HTML content column to the posts table.
 v1.0.3: Category component has been merged with Posts component.
 v1.0.4: Improvements to the Posts list management UI.
@@ -20,15 +21,11 @@ v1.1.0: Replaced the Post editor with the new core Markdown editor.
 v1.1.1: Posts can now be imported and exported.
 v1.1.2: Posts are no longer visible if the published date has not passed.
 v1.1.3: Added a New Post shortcut button to the blog menu.
-v1.2.0:
-    - Categories now support nesting.
-    - categories_add_nested_fields.php
+v1.2.0: Categories now support nesting.
 v1.2.1: Post slugs now must be unique.
 v1.2.2: Fixes issue on new installs.
 v1.2.3: Minor user interface update.
-v1.2.4:
-    - Database maintenance. Updated all timestamp columns to be nullable.
-    - update_timestamp_nullable.php
+v1.2.4: Database maintenance. Updated all timestamp columns to be nullable.
 v1.2.5: Added translation support for blog posts.
 v1.2.6: The published field can now supply a time with the date.
 v1.2.7: Introduced a new RSS feed component.
@@ -44,9 +41,7 @@ v1.2.16: Added `nextPost` and `previousPost` to the blog post component.
 v1.2.17: Improved the next and previous logic to sort by the published date.
 v1.2.18: Minor change to internals.
 v1.2.19: Improved support for Build 420+
-v1.3.0:
-    - Added metadata column for plugins to store data in
-    - posts_add_metadata.php
+v1.3.0: Added metadata column for plugins to store data in.
 v1.3.1: Fixed metadata column not being jsonable
 v1.3.2: Allow custom slug name for components, add 404 handling for missing blog posts, allow exporting of blog images.
 v1.3.3: Fixed 'excluded categories' filter from being run when value is empty.
@@ -71,3 +66,4 @@ v1.6.3: Adds page finder support for October v3.2
 v1.7.0: Adds setting for editing posts with rich editor
 v1.7.1: Fixes bug using Rich Editor in content field
 v1.7.2: Fixes compatibility with PHP 8.3
+v2.0.0: Complete support with October CMS v4.2
